requirements.txt
certifi==2024.6.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
glob2==0.7
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
